requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic>=2.7.0,<3.0.0
pillow==10.4.0
torch>=2.3.0
torchvision>=0.18.0
pymongo[srv]>=4.8.0
bcrypt>=4.1.3
python-dotenv>=1.0.1

